Extracted function to a different file
